requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
